+++ jabgui/src/main/java/org/jabref/gui/maintable/MainTable.java
@@ -581,9 +581,17 @@ public class MainTable extends TableView<BibEntryTableViewModel> {
                         importHandler.importFilesInBackground(files, transferMode).executeWith(taskExecutor);
                 // - Center -> modify entry: link files to entry
                 case CENTER -> {
-                    BibEntry entry = target.getEntry();
-                    ExternalFilesEntryLinker fileLinker = importHandler.getFileLinker();
-                    DragDrop.handleDropOfFiles(files, transferMode, fileLinker, entry);
+                    List<Path> bibFiles = files.stream().filter(FileUtil::isBibFile).toList();
+                    List<Path> otherFiles = files.stream().filter(file -> !FileUtil.isBibFile(file)).toList();
+
+                    if (!bibFiles.isEmpty()) {
+                        importHandler.importFilesInBackground(bibFiles, transferMode).executeWith(taskExecutor);
+                    }
+                    if (!otherFiles.isEmpty()) {
+                        BibEntry entry = target.getEntry();
+                        ExternalFilesEntryLinker fileLinker = importHandler.getFileLinker();
+                        DragDrop.handleDropOfFiles(otherFiles, transferMode, fileLinker, entry);
+                    }
                 }
             }