+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 
 ## [Unreleased]
 
+- Fixed .bib files being moved to the PDF folder when dropped on an entry in the main table. [#16335](https://github.com/JabRef/jabref/issues/16335)
+
 ### Added
 
 - The LibreOffice integration's "Cite special" mode for JStyles can now cite author-year (without parenthesis), author-only and year-only entries. [forum#670](https://discourse.jabref.org/t/cite-special-openoffice-libreoffice-connection-tool/670). [#7861](https://github.com/JabRef/jabref/issues/7861)